first-time-usb-data-stream-on-pico/host/requirements.txt
pyusb
libusb1
matplotlib==3.10.7